% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mult_reg_df}
\alias{mult_reg_df}
\title{Example regression results from multiple models and regressions}
\format{
A tibble with 20 rows and 6 columns
\describe{
\item{variable}{Factor: main variable included  in regression}
\item{estimate}{Numeric: Odds ratio estimate}
\item{width}{Numeric: width of confidence interval}
\item{ub}{Numeric: upper bound of confidence interval}
\item{lb}{Numeric: lower bound of confidence interval}
\item{model}{Character: model that estimate came from}
}
}
\source{
Simulated data, see https://github.com/jhuwit/fine_mapping_sim
}
\usage{
mult_reg_df
}
\description{
A simulated regressions result dataset included in bloodlines to illustrate usage
}
\keyword{datasets}
