% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_j.R
\name{plot_j}
\alias{plot_j}
\title{Create a 'j-plot' from regression results from multiple models}
\usage{
plot_j(
  data,
  facet_var = NULL,
  col_vec = NULL,
  shape_vec = NULL,
  dodge = 0.6,
  point_size = 2,
  bar_width = 0.5,
  line_width = 0.5,
  ybreaks = NULL,
  x_label = "",
  y_label = "",
  title = "",
  ylims = NULL,
  legend_pos = "bottom"
)
}
\arguments{
\item{data}{A data frame containing regression results, must include columns 'variable', 'estimate', 'ub', 'lb', 'model'}

\item{facet_var}{A variable to facet the plot by (optional)}

\item{col_vec}{A vector of colors for the models, may be named (optional)}

\item{shape_vec}{A vector of shapes for the models, may be named (optional)}

\item{dodge}{Amount to dodge the points and error bars (default is 0.6)}

\item{point_size}{Size of the points (default is 2)}

\item{bar_width}{Width of the error bars (default is 0.5)}

\item{line_width}{Width of the error bar lines (default is 0.5)}

\item{ybreaks}{Breaks for the y-axis (optional)}

\item{x_label}{Label for the x-axis (default is "")}

\item{y_label}{Label for the y-axis (default is "")}

\item{title}{Title of the plot (default is "")}

\item{ylims}{Limits for the y-axis (optional, if NULL will be set to min and max of 'lb' and 'ub')}

\item{legend_pos}{position of legend, default is "bottom"}
}
\value{
A 'j-plot'
}
\description{
Create a 'j-plot' from regression results from multiple models
}
\examples{
# Load example walking bout data
data(mult_reg_df)

# Run the function
print(plot_j(mult_reg_df, col_vec = c("A"= "#1E8E99", "B" = "#FF8E32")))

}
