% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts_lasagna.R
\name{plot_ts_lasagna}
\alias{plot_ts_lasagna}
\title{Create a lasagna plot from hemodynamics time series}
\usage{
plot_ts_lasagna(
  data,
  facet_var = NULL,
  layer_var = NULL,
  col_vector = NULL,
  legend_pos = "bottom",
  xlab = "",
  ylab = "",
  title = "",
  xlims = NULL,
  xbreaks = NULL
)
}
\arguments{
\item{data}{A data frame containing hemodynamic time series with columns 'time', 'id', and a variable that is lasagna layers}

\item{facet_var}{A variable to facet the plot by (optional)}

\item{layer_var}{A variable to create lasagna layers from}

\item{col_vector}{A named vector of colors for the layers (default is NULL, which uses default ggplot colors)}

\item{legend_pos}{Position of the legend (default is "bottom")}

\item{xlab}{Label for the x-axis (default is "")}

\item{ylab}{Label for the y-axis (default is "")}

\item{title}{Title of the plot (default is "")}

\item{xlims}{Limits for the x-axis (default is NULL)}

\item{xbreaks}{Breaks for the x-axis (default is NULL)}
}
\value{
A lasagna plot
}
\description{
Create a lasagna plot from hemodynamics time series
}
\examples{
# Load example walking bout data
data(sample_df)

# Run the function
plot_ts_lasagna(sample_df, facet_var = "cat_cpb", layer_var = "map_cat")

}
