% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_df}
\alias{reg_df}
\title{Example regression results}
\format{
A tibble with 25 rows and 7 columns
\describe{
\item{x}{Factor: x-axis variable, e.g. "1" (hemodynamic range)}
\item{y}{Factor: y-axis variable, e.g. "1" (hemodynamic range)}
\item{estimate}{Numeric: Odds ratio estimate}
\item{width}{Numeric: width of confidence interval}
\item{ub}{Numeric: upper bound of confidence interval}
\item{lb}{Numeric: lower bound of confidence interval}
\item{p}{Numeric: p-value for regression}
}
}
\source{
Simulated data, see https://github.com/jhuwit/fine_mapping_sim
}
\usage{
reg_df
}
\description{
A simulated regression result dataset included in bloodlines to illustrate usage
}
\keyword{datasets}
