% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_df}
\alias{sample_df}
\title{Example sample dataset}
\format{
A tibble with 59,008 rows and 13 columns:
\describe{
\item{id}{Character. Subect id, e.g. "1"}
\item{time}{Numeric. Time in minutes from start of recording (by subect)}
\item{timestamp}{POSIXct. Time stamp.}
\item{cat_cpb}{Character. 'pre', 'intra', or 'post' cardiopulmonary bypass (CPB) status.}
\item{val_map}{Numeric. value of mean arterial pressure (MAP) in mmHg}
\item{val_cvp}{Numeric. value of central venous pressure (CVP) in mmHg}
\item{cat_map}{Factor. category of MAP (<65 or >= 65)}
}
}
\source{
Simulated data, see https://github.com/jhuwit/fine_mapping_sim
}
\usage{
sample_df
}
\description{
A simulated hemodynamics dataset included in bloodlines to illustrate usage.
}
\keyword{datasets}
