% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Create a heatmap from regression results where regression predictors can be represented as combination of two variables}
\usage{
plot_heatmap(
  data,
  facet_var = NULL,
  sig = FALSE,
  ci = FALSE,
  show_only_sig = FALSE,
  sig_threshold = 0.05,
  text_col = "white",
  fill_scheme = "gradient2",
  low_col = "blue",
  mid_col = "white",
  high_col = "red",
  mid_value = 1,
  color_vec = NULL,
  n_colors = NULL,
  palette = NULL,
  legend_pos = "bottom",
  fill_lab = "",
  xlab = "",
  ylab = "",
  title = ""
)
}
\arguments{
\item{data}{A data frame containing regression results, must include columns x, y, estimate, p, lb, ub for p-value, upper and lower bounds of CI}

\item{facet_var}{A variable to facet the plot by (optional)}

\item{sig}{Logical indicating whether to show significance (default is FALSE)}

\item{ci}{Logical indicating whether to show confidence intervals (default is FALSE)}

\item{show_only_sig}{Logical indicating whether to show only significant results (default is FALSE)}

\item{sig_threshold}{Threshold for significance (default is 0.05, must be between 0 and 1 if show_only_sig is TRUE)}

\item{text_col}{Color for the text labels (default is white)}

\item{fill_scheme}{One of "gradient", "gradient2", "gradientn", or palette for fill color scheme (default is "gradient2")}

\item{low_col}{Color for the low end of the gradient (default is "blue")}

\item{mid_col}{Color for the middle of the gradient (default is "white"), ONLY for 'gradient2' fill_scheme}

\item{high_col}{Color for the high end of the gradient (default is "red")}

\item{mid_value}{Midpoint value for the 'gradient2' fill scheme (default is 1)}

\item{color_vec}{A vector of colors for the 'gradientn' fill scheme}

\item{n_colors}{Number of colors for the 'gradientn' fill scheme}

\item{palette}{A palette name for the 'palette' fill scheme (default is NULL), e.g.  "viridis::plasma"}

\item{legend_pos}{Position of the legend (default is "bottom")}

\item{fill_lab}{Label for the fill aesthetic (default is "")}

\item{xlab}{Label for the x-axis (default is "")}

\item{ylab}{Label for the y-axis (default is "")}

\item{title}{Title of the plot (default is "")}
}
\value{
A heatmap
}
\description{
Create a heatmap from regression results where regression predictors can be represented as combination of two variables
}
\examples{
# Load example walking bout data
data(reg_df)

# Run the function
plot_heatmap(reg_df)

}
